Move TPraos protocol into its own package.
Intention will be to implement the new Praos protocol also in this
package.
