+++ ouroboros-consensus-cardano/ouroboros-consensus-cardano.cabal
@@ -54,6 +54,7 @@ library
                      , ouroboros-network
                      , ouroboros-consensus
                      , ouroboros-consensus-byron
+                     , ouroboros-consensus-protocol
                      , ouroboros-consensus-shelley
 
   default-language:    Haskell2010
