+++ ouroboros-consensus-protocol/src/Ouroboros/Consensus/Protocol/Ledger/Util.hs
@@ -2,7 +2,7 @@
 --
 -- In particular, various things we need for integration with the @delegation@
 -- package from cardano-ledger-specs.
-module Ouroboros.Consensus.Shelley.Protocol.Util (
+module Ouroboros.Consensus.Protocol.Ledger.Util (
     firstSlotOfEpochOfSlot
   , isNewEpoch
   ) where
